Add files via upload
Our coffee shop final project for the first module. Wanted to try and do some "fancy" things to make it look nice. Pretty happy with how it came out, considering how early on in the class this was! (Images were also generated with Stable Diffusion so that I could be weird)

(Never did get to the food section, but as this was mostly a demonstration of what we could do I decided that would be fine in this particular case as it would have been more of the same code...I'd just have to think of new dumb joke foods to make)